fixed rounding error with blinds factor
git-svn-id: http://svn.code.sf.net/p/holdingnuts/code/trunk@733 00c8e4f2-bc05-11dd-9892-7df5a783e2f7

+++ src/client/pclient.hpp
@@ -103,7 +103,7 @@ typedef struct {
 	//! \brief starting blinds
 	chips_type	blinds_start;
 	//! \brief blinds raise factor
-	float		blinds_factor;
+	double		blinds_factor;
 	//! \brief blinds raise time
 	unsigned int	blinds_time;
 	//! \brief known tables
